# webapps/backend/swagger.json
{"openapi":"3.0.0","paths":{"/MarketSignals":{"get":{"operationId":"MarketSignalsController_getAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MarketSignal"}}}}}},"tags":["MarketSignals"]}},"/MarketSignals/bullish":{"get":{"operationId":"MarketSignalsController_getBullish","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MarketSignal"}}}}}},"tags":["MarketSignals"]}},"/MarketSignals/{symbol}":{"get":{"operationId":"MarketSignalsController_getOne","parameters":[{"name":"symbol","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MarketSignal"}}}}},"tags":["MarketSignals"]}},"/candle-sticks/{symbol}":{"get":{"operationId":"CandleSticksController_findAll","parameters":[{"name":"symbol","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"คืนค่ารายการแท่งเทียน","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CandleStick"}}}}}},"summary":"ดึงข้อมูลแท่งเทียนทั้งหมดตาม Symbol","tags":["CandleSticks"]}},"/candle-sticks/{symbol}/range":{"get":{"operationId":"CandleSticksController_findByRange","parameters":[{"name":"symbol","required":true,"in":"path","schema":{"type":"string"}},{"name":"start","required":true,"in":"query","schema":{"example":"2026-01-01","type":"string"}},{"name":"end","required":true,"in":"query","schema":{"example":"2026-02-23","type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CandleStick"}}}}}},"summary":"ดึงข้อมูลแท่งเทียนตามช่วงวันที่","tags":["CandleSticks"]}},"/candle-sticks":{"post":{"operationId":"CandleSticksController_create","parameters":[],"responses":{"201":{"description":"บันทึกสำเร็จ"}},"summary":"บันทึกข้อมูลแท่งเทียนใหม่","tags":["CandleSticks"]}},"/system-parameters":{"post":{"operationId":"SystemParametersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSystemParameterDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SystemParameter"}}}}},"tags":["SystemParameters"]},"get":{"operationId":"SystemParametersController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SystemParameter"}}}}}},"tags":["SystemParameters"]}},"/system-parameters/{key}/value":{"get":{"operationId":"SystemParametersController_getValue","parameters":[{"name":"key","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["SystemParameters"]}},"/system-parameters/{id}":{"patch":{"operationId":"SystemParametersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSystemParameterDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SystemParameter"}}}}},"tags":["SystemParameters"]}},"/system-permissions":{"post":{"operationId":"SystemPermissionsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSystemPermissionDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SystemPermission"}}}}},"tags":["SystemPermissions"]},"get":{"operationId":"SystemPermissionsController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SystemPermission"}}}}}},"tags":["SystemPermissions"]}},"/system-permissions/{id}":{"get":{"operationId":"SystemPermissionsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SystemPermission"}}}}},"tags":["SystemPermissions"]},"patch":{"operationId":"SystemPermissionsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSystemPermissionDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SystemPermission"}}}}},"tags":["SystemPermissions"]},"delete":{"operationId":"SystemPermissionsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SystemPermission"}}}}},"tags":["SystemPermissions"]}},"/users":{"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["Users"]},"get":{"operationId":"UsersController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}},"tags":["Users"]}},"/users/{id}":{"get":{"operationId":"UsersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["Users"]},"patch":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["Users"]},"delete":{"operationId":"UsersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"]}},"/user-groups":{"post":{"operationId":"UserGroupsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserGroupDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserGroup"}}}}},"tags":["UserGroups"]},"get":{"operationId":"UserGroupsController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserGroup"}}}}}},"tags":["UserGroups"]}},"/user-groups/{id}":{"get":{"operationId":"UserGroupsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserGroup"}}}}},"tags":["UserGroups"]},"patch":{"operationId":"UserGroupsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserGroupDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserGroup"}}}}},"tags":["UserGroups"]},"delete":{"operationId":"UserGroupsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserGroup"}}}}},"tags":["UserGroups"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["Authentication"]}},"/auth/refresh":{"post":{"operationId":"AuthController_refresh","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshDto"}}}},"responses":{"201":{"description":""}},"tags":["Authentication"]}},"/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"201":{"description":""}},"security":[{"accessToken":[]}],"tags":["Authentication"]}},"/auth/logout-all":{"post":{"operationId":"AuthController_logoutAll","parameters":[],"responses":{"201":{"description":""}},"security":[{"accessToken":[]}],"tags":["Authentication"]}}},"info":{"title":"Market Signal API","description":"API สำหรับดึงข้อมูลสัญญาณเทคนิค (RSI, EMA, Crosses)","version":"1.0","contact":{}},"tags":[{"name":"NestJs","description":""}],"servers":[],"components":{"securitySchemes":{"accessToken":{"scheme":"bearer","bearerFormat":"JWT","type":"http","name":"JWT","description":"ใส่เฉพาะ JWT Token (ไม่ต้องพิมพ์ Bearer นำหน้า)","in":"header"}},"schemas":{"MarketSignal":{"type":"object","properties":{"id":{"type":"number"},"symbol":{"type":"string"},"last_close":{"type":"number"},"signal":{"type":"string"},"trend":{"type":"string"},"rsi_14":{"type":"number"},"rsi_overbought":{"type":"number"},"rsi_oversold":{"type":"number"},"ema_50":{"type":"number"},"ema_200":{"type":"number"},"golden_cross":{"type":"boolean"},"death_cross":{"type":"boolean"},"updated_at":{"format":"date-time","type":"string"}},"required":["id","symbol","last_close","signal","trend","rsi_14","rsi_overbought","rsi_oversold","ema_50","ema_200","golden_cross","death_cross","updated_at"]},"CandleStick":{"type":"object","properties":{"symbol":{"type":"string"},"x":{"format":"date-time","type":"string"},"o":{"type":"number"},"h":{"type":"number"},"l":{"type":"number"},"c":{"type":"number"},"v":{"type":"number"}},"required":["symbol","x","o","h","l","c","v"]},"CreateSystemParameterDto":{"type":"object","properties":{"paramKey":{"type":"string"},"paramValue":{"type":"string"},"valueType":{"type":"string"},"description":{"type":"string"}},"required":["paramKey","paramValue"]},"SystemParameter":{"type":"object","properties":{"paramId":{"type":"number"},"paramKey":{"type":"string"},"paramValue":{"type":"string"},"valueType":{"type":"string"},"description":{"type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["paramId","paramKey","paramValue","valueType","description","updatedAt"]},"UpdateSystemParameterDto":{"type":"object","properties":{}},"CreateSystemPermissionDto":{"type":"object","properties":{"permissionKey":{"type":"string","pattern":"/^[a-z0-9.]*$/"},"permissionName":{"type":"string"},"description":{"type":"string"}},"required":["permissionKey","permissionName"]},"SystemPermission":{"type":"object","properties":{"permissionId":{"type":"number"},"permissionKey":{"type":"string"},"permissionName":{"type":"string"},"description":{"type":"string"}},"required":["permissionId","permissionKey","permissionName","description"]},"UpdateSystemPermissionDto":{"type":"object","properties":{}},"CreateUserDto":{"type":"object","properties":{"username":{"type":"string"},"email":{"type":"string","format":"email"},"passwordHash":{"type":"string","minLength":8},"groupId":{"type":"number"},"status":{"type":"string"}},"required":["username","email","passwordHash","groupId"]},"User":{"type":"object","properties":{"userId":{"type":"string"},"username":{"type":"string"},"email":{"type":"string"},"passwordHash":{"type":"string"},"group":{"$ref":"#/components/schemas/UserGroup"},"groupId":{"type":"number"},"isActive":{"type":"boolean"},"status":{"type":"string"},"isLoggedIn":{"type":"boolean"},"sessionKey":{"type":"string"},"loginAttempts":{"type":"number"},"lockUntil":{"format":"date-time","type":"string"},"lastLogin":{"format":"date-time","type":"string"},"createdAt":{"format":"date-time","type":"string"},"createdBy":{"type":"string"},"updatedAt":{"format":"date-time","type":"string"},"updatedBy":{"type":"string"}},"required":["userId","username","email","passwordHash","group","groupId","isActive","status","isLoggedIn","sessionKey","loginAttempts","lockUntil","lastLogin","createdAt","createdBy","updatedAt","updatedBy"]},"UserGroup":{"type":"object","properties":{"groupId":{"type":"number"},"groupName":{"type":"string"},"description":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"users":{"type":"array","items":{"$ref":"#/components/schemas/User"}},"permissions":{"type":"array","items":{"$ref":"#/components/schemas/SystemPermission"}}},"required":["groupId","groupName","description","createdAt","users","permissions"]},"UpdateUserDto":{"type":"object","properties":{"isActive":{"type":"boolean"},"isLoggedIn":{"type":"boolean"}}},"CreateUserGroupDto":{"type":"object","properties":{"groupName":{"type":"string"},"description":{"type":"string"},"permissionIds":{"type":"array","items":{"type":"number"}}},"required":["groupName"]},"UpdateUserGroupDto":{"type":"object","properties":{}},"LoginDto":{"type":"object","properties":{"username":{"type":"string","description":"The username of the user","example":"tester01"},"password":{"type":"string","description":"The password of the user","example":"SecretPassword123"}},"required":["username","password"]},"RefreshDto":{"type":"object","properties":{"refreshToken":{"type":"string"}},"required":["refreshToken"]}}}}